refactor(templates): centralize ownership map
