--- plugins/prompt-templates/comment-polish/ownership.mjs
@@ -1,5 +1,0 @@
-export const commentPolishOwnership = {
-    templateId: 'builtin_comment_polish',
-    createdBy: 'ymkiux',
-    maintainers: ['ymkiux']
-};
