+++ plugins/prompt-templates/ownership.mjs
@@ -1,0 +1,19 @@
+export const pluginOwnership = {
+    pluginId: 'prompt-templates',
+    createdBy: 'ymkiux',
+    maintainers: ['ymkiux']
+};
+
+export const templateOwnershipById = {
+    builtin_comment_polish: {
+        templateId: 'builtin_comment_polish',
+        createdBy: 'ymkiux',
+        maintainers: ['ymkiux']
+    },
+    builtin_rule_ack: {
+        templateId: 'builtin_rule_ack',
+        createdBy: 'ymkiux',
+        maintainers: ['ymkiux']
+    }
+};
+
