--- plugins/prompt-templates/plugin-ownership.mjs
@@ -1,6 +1,0 @@
-export const pluginOwnership = {
-    pluginId: 'prompt-templates',
-    createdBy: 'ymkiux',
-    maintainers: ['ymkiux']
-};
-
