--- plugins/prompt-templates/rule-ack/ownership.mjs
@@ -1,5 +1,0 @@
-export const ruleAckOwnership = {
-    templateId: 'builtin_rule_ack',
-    createdBy: 'ymkiux',
-    maintainers: ['ymkiux']
-};
